requirements.txt
# Core Django
Django==5.0.1
psycopg2-binary==2.9.9

# Authentication
django-allauth==0.57.0

# Frontend & UI
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

# Markdown support
django-markdownx==4.0.7
markdown==3.5.2
bleach==6.1.0

# Rate limiting
django-ratelimit==4.1.0

# Image handling
Pillow==10.2.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0
dj-database-url==2.1.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Fuzzy search for SQLite
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
